implement idea of `alias_fqdn`
With this commit, the structure of the `inventory.yml` file based on server aliases was improved, so we can use the `ansible_host` defined there as default value for hostname and URL retrievals and fall back on the new created `alias_fqdn` variable i.e. if the `alias_host` simply is an IP address or sth. else, we do not want to be represented within the hostname.